# openapi.yaml
...
openapi: "3.0.3"
info:
  title: "identuum-idp-oss API"
  version: "identuum-idp-oss-api-docgen 0.2.0-p3"
  description: "Machine-generated OpenAPI 3 specification for the identuum-idp-oss API surface. Generated by tools/api-docgen from // docgen: source annotations. Response bodies without a known schema record are described but not field-modeled; see the referenced Go symbol in identuum-idp-oss for the authoritative shape."
paths:
...
  "/api/v1/organizations/{id}/service-accounts":
    get:
      operationId: "identuum-idp-oss.service-accounts.list_service_accounts"
      summary: "List service accounts in an organization (handler enforces site_admin OR same-org membership)."
      tags:
        - "service-accounts"
      parameters:
        - name: "id"
          in: path
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "200":
          description: "List service accounts in an organization (handler enforces site_admin OR same-org membership)."
        default:
          description: "Error response — see identuum-idp-oss source for exact status/body per failure mode."
    post:
      operationId: "identuum-idp-oss.service-accounts.create_service_account"
      summary: "Create a service account in an organization (handler enforces site_admin OR same-org membership)."
      tags:
        - "service-accounts"
      parameters:
        - name: "id"
          in: path
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "201":
          description: "Create a service account in an organization (handler enforces site_admin OR same-org membership)."
        default:
          description: "Error response — see identuum-idp-oss source for exact status/body per failure mode."
  "/api/v1/organizations/{id}/service-accounts/with-client":
    post:
      operationId: "identuum-idp-oss.service-accounts.create_service_account_with_client"
      summary: "Create a service account bundled with an OAuth client (one-shot issuance; the OAuth client_secret is returned ONCE)."
      tags:
        - "service-accounts"
      parameters:
        - name: "id"
          in: path
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "201":
          description: "Create a service account bundled with an OAuth client (one-shot issuance; the OAuth client_secret is returned ONCE)."
        default:
          description: "Error response — see identuum-idp-oss source for exact status/body per failure mode."
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: "identuum_session"
      description: "Browser session cookie backing a cookie-authenticated Principal (site_admin, org_admin, or any authenticated user, depending on the operation's actual role requirement — see its summary)."
    oauthClientAuth:
      type: http
      scheme: basic
      description: "OAuth 2.1 client authentication (client credentials / RFC 7592 management token) at client-management and token endpoints."
...